+++ test/test_test.exs
@@ -213,4 +213,91 @@ defmodule Maru.TestTest do
     assert false == Map.has_key?(private, :maru_plug_test2)
   end
 
+  test "mounted with exception handlers" do
+    defmodule MWEH do
+      use Maru.Router
+      mount Maru.TestTest.MWEH.M1
+
+      rescue_from :all do
+        conn |> put_status(500) |> text("500")
+      end
+    end
+
+    defmodule MWEH.M1 do
+      use Maru.Router
+      mount Maru.TestTest.MWEH.M2
+
+      rescue_from ArithmeticError do
+        conn |> put_status(501) |> text("501")
+      end
+    end
+
+    defmodule MWEH.M2 do
+      use Maru.Router
+
+      rescue_from MatchError do
+        conn |> put_status(502) |> text("502")
+      end
+
+      get do
+        text(conn, "200")
+      end
+
+      get "match_error" do
+        1 = 2
+        text(conn, "200")
+      end
+
+      get "arithmetic_error" do
+        2 / 0
+        text(conn, "200")
+      end
+
+      get "runtime_error" do
+        raise "runtime_error"
+        text(conn, "200")
+      end
+    end
+
+    for module <- [Maru.TestTest.MWEH, Maru.TestTest.MWEH.M1, Maru.TestTest.MWEH.M2] do
+      Enum.each(module.__mounted_modules__, fn mounted ->
+        Maru.Builder.MountLink.put_father(mounted, module)
+      end)
+    end
+
+    defmodule MWEH.M2.TEST1 do
+      use Maru.Test, for: Maru.TestTest.MWEH.M2
+
+      def test1, do: get("/").status
+      def test2, do: get("/match_error").status
+      def test3, do: get("/arithmetic_error").status
+      def test4, do: get("/runtime_error").status
+    end
+
+    assert 200 == MWEH.M2.TEST1.test1
+    assert 502 == MWEH.M2.TEST1.test2
+    assert_raise ArithmeticError, fn ->
+      MWEH.M2.TEST1.test3
+    end
+    assert_raise RuntimeError, fn ->
+      MWEH.M2.TEST1.test4
+    end
+
+    defmodule MWEH.M2.TEST2 do
+      use Maru.Test,
+        for: Maru.TestTest.MWEH.M2,
+        with_exception_handlers: true
+
+      def test1, do: get("/").status
+      def test2, do: get("/match_error").status
+      def test3, do: get("/arithmetic_error").status
+      def test4, do: get("/runtime_error").status
+    end
+
+    assert 200 == MWEH.M2.TEST2.test1
+    assert 502 == MWEH.M2.TEST2.test2
+    assert 501 == MWEH.M2.TEST2.test3
+    assert 500 == MWEH.M2.TEST2.test4
+  end
+
 end